@@ -16,63 +16,64 @@ namespace toTwig;
  */
 interface ConfigInterface
 {
-	/**
-	 * Returns the name of the configuration.
-	 *
-	 * The name must be all lowercase and without any spaces.
-	 *
-	 * @return string The name of the configuration
-	 */
-	public function getName();
 
-	/**
-	 * Returns the description of the configuration.
-	 *
-	 * A short one-line description for the configuration.
-	 *
-	 * @return string The description of the configuration
-	 */
-	public function getDescription();
+    /**
+     * Returns the name of the configuration.
+     *
+     * The name must be all lowercase and without any spaces.
+     *
+     * @return string The name of the configuration
+     */
+    public function getName();
 
-	/**
-	 * Returns an iterator of files to scan.
-	 *
-	 * @return \Traversable A \Traversable instance that returns \SplFileInfo instances
-	 */
-	public function getFinder();
+    /**
+     * Returns the description of the configuration.
+     *
+     * A short one-line description for the configuration.
+     *
+     * @return string The description of the configuration
+     */
+    public function getDescription();
 
-	/**
-	 * Returns the converters to run.
-	 *
-	 * @return array|integer A level or a list of converter names
-	 */
-	public function getConverters();
+    /**
+     * Returns an iterator of files to scan.
+     *
+     * @return \Traversable A \Traversable instance that returns \SplFileInfo instances
+     */
+    public function getFinder();
 
-	/**
-	 * Sets the root directory of the project.
-	 *
-	 * @param string $dir The project root directory
-	 */
-	public function setDir($dir);
+    /**
+     * Returns the converters to run.
+     *
+     * @return array|integer A level or a list of converter names
+     */
+    public function getConverters();
 
-	/**
-	 * Returns the root directory of the project.
-	 *
-	 * @return string The project root directory
-	 */
-	public function getDir();
+    /**
+     * Sets the root directory of the project.
+     *
+     * @param string $dir The project root directory
+     */
+    public function setDir($dir);
 
-	/**
-	 * Adds an instance of a custom converter.
-	 *
-	 * @param ConverterInterface $converter
-	 */
-	public function addCustomConverter(ConverterAbstract $converter);
+    /**
+     * Returns the root directory of the project.
+     *
+     * @return string The project root directory
+     */
+    public function getDir();
 
-	/**
-	 * Returns the custom converters to use.
-	 *
-	 * @return ConverterInterface[]
-	 */
-	public function getCustomConverters();
+    /**
+     * Adds an instance of a custom converter.
+     *
+     * @param ConverterAbstract $converter
+     */
+    public function addCustomConverter(ConverterAbstract $converter);
+
+    /**
+     * Returns the custom converters to use.
+     *
+     * @return ConverterAbstract[]
+     */
+    public function getCustomConverters();
 }
