@@ -20,23 +20,26 @@ use Symfony\Component\Console\Output\OutputInterface;
  */
 class ReadmeCommand extends Command
 {
-	/**
-	 * @see Command
-	 */
-	protected function configure()
-	{
-		$this
-			->setName('readme')
-			->setDescription('Generates the README content, based on the convert command help')
-		;
-	}
 
-	/**
-	 * @see Command
-	 */
-	protected function execute(InputInterface $input, OutputInterface $output)
-	{
-		$header = <<<EOF
+    /**
+     * @see Command
+     */
+    protected function configure()
+    {
+        $this
+            ->setName('readme')
+            ->setDescription('Generates the README content, based on the convert command help');
+    }
+
+    /**
+     * @param InputInterface  $input
+     * @param OutputInterface $output
+     *
+     * @see Command
+     */
+    protected function execute(InputInterface $input, OutputInterface $output)
+    {
+        $header = <<<EOF
 PHP Smarty to Twig Converter
 ==========================
 toTwig is an utility to convert smarty template engine to twig template engine.
@@ -86,7 +89,7 @@ Usage
 
 EOF;
 
-		$footer = <<<EOF
+        $footer = <<<EOF
 
 Contribute
 ----------
@@ -109,22 +112,22 @@ projects for instance).
 
 EOF;
 
-		$command = $this->getApplication()->get('convert');
-		$help = $command->getHelp();
-		$help = str_replace('%command.full_name%', 'toTwig.phar '.$command->getName(), $help);
-		$help = str_replace('%command.name%', $command->getName(), $help);
-		$help = preg_replace('#</?(comment|info)>#', '`', $help);
-		$help = preg_replace('#^(\s+)`(.+)`$#m', '$1$2', $help);
-		$help = preg_replace('#^ \* `(.+)`#m', ' * **$1**', $help);
+        $command = $this->getApplication()->get('convert');
+        $help = $command->getHelp();
+        $help = str_replace('%command.full_name%', 'toTwig.phar ' . $command->getName(), $help);
+        $help = str_replace('%command.name%', $command->getName(), $help);
+        $help = preg_replace('#</?(comment|info)>#', '`', $help);
+        $help = preg_replace('#^(\s+)`(.+)`$#m', '$1$2', $help);
+        $help = preg_replace('#^ \* `(.+)`#m', ' * **$1**', $help);
 
-		$output->writeln($header);
-		$output->writeln($help);
-		$output->write($footer);
+        $output->writeln($header);
+        $output->writeln($help);
+        $output->write($footer);
 
-		$content = $header."\n";
-		$content .= $help ."\n";
-		$content .= $footer;
+        $content = $header . "\n";
+        $content .= $help . "\n";
+        $content .= $footer;
 
-		@file_put_contents('README.md', $content);
-	}
+        @file_put_contents('README.md', $content);
+    }
 }
