@@ -18,98 +18,115 @@ use toTwig\ConverterAbstract;
  */
 class ForConverter extends ConverterAbstract
 {
-	// Lookup tables for performing some token
-	// replacements not addressed in the grammar.
-	private $replacements = array(
-		'smarty\.foreach.*\.index' => 'loop.index0',
-		'smarty\.foreach.*\.iteration' => 'loop.index'
-	);
 
-	public function convert(\SplFileInfo $file, $content)
-	{
-		$content = $this->replaceFor($content);
-		$content = $this->replaceEndForEach($content);
-		$content = $this->replaceForEachElse($content);
+    protected $name = 'for';
+    protected $description = 'Convert foreach/foreachelse to twig';
+    protected $priority = 50;
 
-		foreach ($this->replacements as $k=>$v) {
-			$content = preg_replace('/'.$k.'/', $v, $content);
-		}
+    // Lookup tables for performing some token
+    // replacements not addressed in the grammar.
+    private $replacements = array(
+        'smarty\.foreach.*\.index'     => 'loop.index0',
+        'smarty\.foreach.*\.iteration' => 'loop.index'
+    );
 
-		return $content;
-	}
+    /**
+     * @param \SplFileInfo $file
+     * @param string       $content
+     *
+     * @return string
+     */
+    public function convert(\SplFileInfo $file, $content)
+    {
+        $content = $this->replaceFor($content);
+        $content = $this->replaceEndForEach($content);
+        $content = $this->replaceForEachElse($content);
 
-	public function getPriority()
-	{
-		return 50;
-	}
+        foreach ($this->replacements as $k => $v) {
+            $content = preg_replace('/' . $k . '/', $v, $content);
+        }
 
-	public function getName()
-	{
-		return 'for';
-	}
+        return $content;
+    }
 
-	public function getDescription()
-	{
-		return 'Convert foreach/foreachelse to twig';
-	}
+    /**
+     * @param string $content
+     *
+     * @return string
+     */
+    private function replaceEndForEach($content)
+    {
+        // [{/foreach}]
+        $search = $this->getClosingTagPattern('foreach');
+        $replace = "{% endfor %}";
 
-	private function replaceEndForEach($content)
-	{
-		$search = "#\{/foreach\s*\}#";
-		$replace = "{% endfor %}";
-		return preg_replace($search,$replace,$content);
-	}
+        return preg_replace($search, $replace, $content);
+    }
 
-	private function replaceForEachElse($content)
-	{
-		$search = "#\{foreachelse\s*\}#";
-		$replace = "{% else %}";
-		return preg_replace($search,$replace,$content);
-	}
+    /**
+     * @param string $content
+     *
+     * @return string
+     */
+    private function replaceForEachElse($content)
+    {
+        // [{foreachelse other stuff}]
+        $search = $this->getOpeningTagPattern('foreachelse');
+        $replace = "{% else %}";
 
-	private function replaceFor($content){
+        return preg_replace($search, $replace, $content);
+    }
 
-		// $pattern = "#\{foreach\b\s*(?:(?!}).)+?\}#";
-		$pattern = "#\{foreach\b\s*([^{}]+)?\}#i";
-		$string  = '{% for :key :item in :from %}';
+    /**
+     * @param string $content
+     *
+     * @return string
+     */
+    private function replaceFor($content)
+    {
+        // [{foreach other stuff}]
+        $pattern = $this->getOpeningTagPattern('foreach');
+        $string = '{% for :key :item in :from %}';
 
-		return preg_replace_callback($pattern, function($matches) use( $string ) {
+        return preg_replace_callback(
+            $pattern,
+            function ($matches) use ($string) {
+                $match = $matches[1];
+                $search = $matches[0];
+                $replace = [];
 
-			$match   = $matches[1];
-			$search  = $matches[0];
-			$replace = array();
+                // {foreach $users as $user}
+                if (preg_match("/(.*)(?:\bas\b)(.*)/i", $match, $mcs)) {
+                    // {foreach $users as $k => $val}
+                    if (preg_match("/(.*)\=\>(.*)/", $mcs[2], $match)) {
+                        if (!isset($replace['key'])) {
+                            $replace['key'] = '';
+                        }
+                        $replace['key'] .= $this->variable($match[1]) . ',';
+                        $mcs[2] = $match[2];
+                    }
+                    $replace['item'] = $this->variable($mcs[2]);
+                    $replace['from'] = $mcs[1];
+                } else {
+                    $attr = $this->attributes($match);
 
-			// {foreach $users as $user}
-			if (preg_match("/(.*)(?:as)(.*)/i", $match,$mcs)) {
+                    if (isset($attr['key'])) {
+                        $replace['key'] = $attr['key'] . ',';
+                    }
 
-				// {foreach $users as $k => $val}
-				if (preg_match("/(.*)\=\>(.*)/", $mcs[2],$match)) {
-					$replace['key'] .= $this->variable($match[1]).',';
-					$mcs[2] = $match[2];
-				} 
-				$replace['item'] = $this->variable($mcs[2]);
-				$replace['from'] = $this->variable($mcs[1]);
+                    $replace['item'] = $this->variable($attr['item']);
+                    $replace['from'] = $attr['from'];
+                }
 
-			} else {
+                $replace['from'] = $this->value($replace['from']);
 
-				$attr = $this->attributes($match);
+                $string = $this->vsprintf($string, $replace);
+                // Replace more than one space to single space
+                $string = preg_replace('!\s+!', ' ', $string);
 
-				if ($attr['key']) {
-					$replace['key'] = $attr['key'].',';
-				}
-
-				$replace['item'] = $this->variable($attr['item']);
-				$replace['from'] = $this->variable($attr['from']);
-			}
-
-			$string  = $this->vsprintf($string,$replace);
-	        // Replace more than one space to single space
-	        $string = preg_replace('!\s+!', ' ', $string);
-
-	        return str_replace($search, $string, $search);
-
-		}, $content);
-
-	}
-
+                return str_replace($search, $string, $search);
+            },
+            $content
+        );
+    }
 }
